+++ src/error.rs
@@ -21,12 +21,20 @@ pub enum StatsError {
     ArgIntervalExclMax(&'static str, f64, f64),
     /// An argument must have been greater than a value but wasn't
     ArgGt(&'static str, f64),
+    /// An argument must have been greater than another argument but wasn't
+    ArgGtArg(&'static str, &'static str),
     /// An argument must have been greater than or equal to a value but wasn't
     ArgGte(&'static str, f64),
+    /// An argument must have been greater than or equal to another argument but wasn't
+    ArgGteArg(&'static str, &'static str),
     /// An argument must have been less than a value but wasn't
     ArgLt(&'static str, f64),
+    /// An argument must have been less than another argument but wasn't
+    ArgLtArg(&'static str, &'static str),
     /// An argument must have been less than or equal to a value but wasn't
     ArgLte(&'static str, f64),
+    /// An argument must have been less than or equal to another argument but wasn't
+    ArgLteArg(&'static str, &'static str),
     /// Containers of the same length were expected
     ContainersMustBeSameLength,
     /// Computation failed to converge,
@@ -62,11 +70,21 @@ impl fmt::Display for StatsError {
                 write!(f, "Argument {} not within interval [{}, {})", s, min, max)
             }
             StatsError::ArgGt(s, val) => write!(f, "Argument {} must be greater than {}", s, val),
+            StatsError::ArgGtArg(s, val) => {
+                write!(f, "Argument {} must be greater than {}", s, val)
+            }
             StatsError::ArgGte(s, val) => {
                 write!(f, "Argument {} must be greater than or equal to {}", s, val)
             }
+            StatsError::ArgGteArg(s, val) => {
+                write!(f, "Argument {} must be greater than or equal to {}", s, val)
+            }
             StatsError::ArgLt(s, val) => write!(f, "Argument {} must be less than {}", s, val),
+            StatsError::ArgLtArg(s, val) => write!(f, "Argument {} must be less than {}", s, val),
             StatsError::ArgLte(s, val) => {
+                write!(f, "Argument {} must be less than or equal to {}", s, val)
+            }
+            StatsError::ArgLteArg(s, val) => {
                 write!(f, "Argument {} must be less than or equal to {}", s, val)
             }
             StatsError::ContainersMustBeSameLength => {