+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.5.0
  - Implemented the `harmonic::gen_harmonic` special function
  - Implemented the `InverseGamma` distribution
  - Implemented the `Geometric` distribution
+ - Implemented the `Hypergeometric ` distribution
  - `gamma::gamma_ur` now panics when `x > 0` or `a == f64::NEG_INFINITY`. In addition, it also returns `f64::NAN` when `a == f64::INFINITY` and `0.0` when `x == f64::INFINITY`
  - `Gamma::pdf` and `Gamma::ln_pdf` now return `f64::NAN` if any of `shape`, `rate`, or `x` are `f64::INFINITY`
  - `Binomial::pdf` and `Binomial::ln_pdf` now panic if `x > n` or `x < 0`